requirements-travis.txt
Django
gunicorn
django-heroku
django-allauth
django-embed_video
